pyproject.toml
[project]
name = "azure-sql-db-tools"
version = "0.1.0"
description = "A CLI tool for querying Azure SQL Database logs and diagnostics from Log Analytics workspaces, providing a terminal-based alternative to the Azure Portal interface."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-monitor-query",
    "azure-identity",
    "rich>=14.0.0",
]

[project.scripts]
azure-sql-db-tools = "azure_sql_db_tools:main"
db-errors = "azure_sql_db_tools.cli:errors"
db-slow-queries = "azure_sql_db_tools.cli:slow_queries"

[project.optional-dependencies]
dev = ["ruff", "mypy"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "./src" }

[tool.ruff]
target-version = "py313"
indent-width = 4
line-length = 100
exclude = ["*.json"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "W", "C90"]

[tool.ruff.format]
indent-style = "space"
line-ending = "auto"
quote-style = "double"

[tool.mypy]
mypy_path = "src"
disallow_untyped_defs = true
python_version = "3.13"
warn_return_any = true
strict = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
explicit_package_bases = true
namespace_packages = true

[tool.pyright]
venvPath = "."
venv = ".venv"

[dependency-groups]
dev = [
    "mypy>=1.16.1",
]
